feat: update .gitignore to include new configuration files
