+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "podscape-electron",
-  "version": "3.1.3",
+  "version": "3.2.0",
   "description": "Kubernetes management desktop app — Electron + React + TypeScript",
   "main": "./out/main/index.js",
   "author": "Coding Protocols Private Limited <support@codingprotocols.com>",
