+++ src/renderer/components/core/SectionRouter.tsx
@@ -5,6 +5,8 @@ import { Search, Terminal } from 'lucide-react'
 import { RefreshButton } from '../common'
 import { LIST_SECTIONS, CLUSTER_SCOPED_SECTIONS, SECTION_LABELS, PROVIDER_SECTIONS } from '../../config'
 import { KubeCRD, ResourceKind } from '../../types'
+import type { CreatableKind } from '../common/CreateResourceModal'
+import { canVerb } from '../../store/slices/clusterSlice'
 
 import Dashboard from './Dashboard'
 import ResourceList from './ResourceList'
@@ -31,6 +33,14 @@ const DebugPodLauncher = React.lazy(() => import('../advanced/DebugPodLauncher')
 const ProviderResourcePanel = React.lazy(() => import('../panels/ProviderResourcePanel'))
 const CostPanel = React.lazy(() => import('../panels/CostPanel'))
 const KrewPanel = React.lazy(() => import('../panels/KrewPanel'))
+
+const CREATABLE_SECTIONS: Partial<Record<string, CreatableKind>> = {
+  deployments: 'deployment',
+  services: 'service',
+  configmaps: 'configmap',
+  secrets: 'secret',
+  namespaces: 'namespace',
+}
 
 // Sections that render directly (no Suspense needed)
 const DIRECT_PANELS: Partial<Record<ResourceKind, React.ComponentType>> = {
@@ -68,6 +78,8 @@ export default function SectionRouter(): JSX.Element {
     kubeconfigOk,
     selectResource,
     execSessions,
+    allowedVerbs,
+    openCreate,
   } = useAppStore(useShallow(s => ({
     section: s.section,
     selectedNamespace: s.selectedNamespace,
@@ -79,11 +91,15 @@ export default function SectionRouter(): JSX.Element {
     kubeconfigOk: s.kubeconfigOk,
     selectResource: s.selectResource,
     execSessions: s.execSessions,
+    allowedVerbs: s.allowedVerbs,
+    openCreate: s.openCreate,
   })))
 
   if (!kubeconfigOk) {
     return <KubeConfigOnboarding />
   }
+
+  const creatableKind = CREATABLE_SECTIONS[section]
 
   const DirectPanel = DIRECT_PANELS[section]
   const LazyPanel = LAZY_PANELS[section]
@@ -148,6 +164,21 @@ export default function SectionRouter(): JSX.Element {
                 loading={loadingResources}
                 title="Refresh"
               />
+
+              {creatableKind && canVerb(allowedVerbs, section, 'create') && (
+                <button
+                  type="button"
+                  onClick={() => openCreate(creatableKind)}
+                  className="flex items-center gap-1.5 px-3 py-2 text-[11px] font-bold rounded-xl
+                             bg-blue-600 hover:bg-blue-500 text-white transition-colors
+                             shadow-sm shadow-blue-500/30"
+                >
+                  <svg width="10" height="10" viewBox="0 0 10 10" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round">
+                    <path d="M5 1v8M1 5h8" />
+                  </svg>
+                  New
+                </button>
+              )}
             </div>
           </PageHeader>
